+++ tests/output/json_line.py
@@ -13,7 +13,6 @@ from plaso.formatters import manager as formatters_manager
 from plaso.lib import timelib
 from plaso.output import json_line
 
-from tests.cli import test_lib as cli_test_lib
 from tests.output import test_lib
 
 
@@ -23,7 +22,7 @@ class JSONLinesOutputTest(test_lib.OutputModuleTestCase):
   def setUp(self):
     """Makes preparations before running an individual test."""
     output_mediator = self._CreateOutputMediator()
-    self._output_writer = cli_test_lib.TestOutputWriter()
+    self._output_writer = test_lib.TestOutputWriter()
     self._output_module = json_line.JSONLineOutputModule(output_mediator)
     self._output_module.SetOutputWriter(self._output_writer)
     self._event_object = test_lib.TestEventObject()
