+++ tests/output/json_out.py
@@ -12,7 +12,6 @@ import unittest
 from plaso.lib import timelib
 from plaso.output import json_out
 
-from tests.cli import test_lib as cli_test_lib
 from tests.output import test_lib
 
 
@@ -22,7 +21,7 @@ class JSONOutputTest(test_lib.OutputModuleTestCase):
   def setUp(self):
     """Makes preparations before running an individual test."""
     output_mediator = self._CreateOutputMediator()
-    self._output_writer = cli_test_lib.TestOutputWriter()
+    self._output_writer = test_lib.TestOutputWriter()
     self._output_module = json_out.JSONOutputModule(output_mediator)
     self._output_module.SetOutputWriter(self._output_writer)
     self._event_object = test_lib.TestEventObject()
